+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed issue in `Trajectory.read_pivot()` where timestamps were not correctly extracted from actors
 - Fixed bug in `Trajectory.__len__()` method for single-point trajectories
+- Added string representation methods (`__str__`) to `Cartesian3` and `Cartographic` classes to avoid errors when printing instances of these classes
 - Corrected minor typos in the `Trajectory` class documentation strings
 
 ## [0.4.0] - 2025-10-23